refactor(SwingSet): consolidate kernel queue logic
